Create subfolder for binary utilities
